repository: implement redis repo backend
This introduces a new RedisRepository in addition to the original Django
ORM-based one. The app will use redis if the `GBP_PS_REDIS_URL`
environment variable is set. Otherwise it uses the original Django one.

The reason why Django is the default is because it's easier to install.
Django is already installed with GBP and adding gbp-ps and going a db
migration makes it work easily. For redis you need to set up an
additional redis server, etc. I personally think the redis backend is
better though.