backend: use gentoo-build-publisher Settings
The gentoo-build-publisher package gained a BaseSettings class that we
can utilize instead of checking for environment variables ourselves.
This change utilizes that class.

This also means that gbp-ps now explicitly imports from
gentoo_build_publisher, so I had to put a hard dependency on
gentoo_build_publisher. However in doing so I created a new group,
"backend", so that users only wanting to use the command-line interface
don't have to pull in all of gentoo_build_publisher.  I also added a
group for redis as the default backend is django.