add signal handlers for gbp prepull and postpull
This signal handlers will add a process entry. The package will be
"pipeline" and the phase will be "pull" for the "prepull" handler and
"clean" for the "postpull" handler.

This will enable gbp-ps to show GBP "pull" phases as an "ebuild"
process but it's really a pipeline process which is why the package is
called "pipeline".