repository: don't let build1's index wipe out build2's world
Take the following scenario:

| t | machine | build | package  | phase |
|---|---------|-------|----------|-------|
| 1 | babette |     1 | pipeline | clean |
| 2 | babette |     2 | pipeline | world |
| 3 | babette |     1 | pipeline | index |

In this case at time t3 the indexing of the previous build is wiping out
the process entry for the current build (build 2).  We don't want that
to happen unless it's an actual package building. In other world unless
the process's phases is in BuildProcess.build_phases.