signals: create signal handler factory function
All of the handlers are basically the same. We can write a factory
function to return the handlers. Unfortunately pydispatch will garbage
collect handlers that fall out of reference, so we need an list to hold
a reference to the handlers. Also we use typing.Protocol because
typing.Callable doesn't seem to be able to handle the wonky handler
signature.